refactor(webapp-game): move off remove and refill jobs move to actionMoves
